corrcted minor bug of size and changed the mode of saving of dilated files to save space

+++ prune_skel.py
@@ -56,7 +56,7 @@ import sys
 
 plate = int(sys.argv[1])
 i = int(sys.argv[-1])
-threshold = int(sys.argv[2])
+threshold = float(sys.argv[2])
 from directory import directory
 
 dates_datetime = get_dates_datetime(directory,plate)
